requirements.txt
streamlit==1.39.0
pillow>=11.0.0
python-dateutil==2.9.0.post0
